requirements.txt
Django>=1.3
django-uni-form
